Added "Ammo" columns to the "Assign" tab
It's similar to the drug policies, but easier to access and only allows carrying one type of ammo, which is why I didn't remove the drug policy option.
Only added two columns for ammo, but a third one can easily be added by uncommenting the "yy_ammo3" line in patch_pawntabledef!